move speculative_execution.rs to new execution mod
