+++ scylla/src/execution/mod.rs
@@ -3,9 +3,11 @@
 //! - execution profiles,
 //! - policies:
 //!     - load balancing,
+//!     - speculative execution,
 //! - TODO.
 
 pub mod execution_profile;
 pub mod load_balancing;
+pub mod speculative_execution;
 
 pub use execution_profile::ExecutionProfile;
