+++ scylla/src/transport/mod.rs
@@ -7,7 +7,6 @@ pub(crate) mod metrics;
 pub mod partitioner;
 pub mod query_result;
 pub mod retry_policy;
-pub mod speculative_execution;
 
 pub use crate::frame::Authenticator;
 pub use scylla_cql::frame::request::query::{PagingState, PagingStateResponse};
